+++ android-generator/src/main/scala/models/generator/javapojos/AndroidRxClasses.scala
@@ -1,4 +1,4 @@
-package models.generator.android
+package models.generator.javapojos
 
 import com.squareup.javapoet.ClassName
 import io.reactivex.Single
