+++ android-generator/src/main/scala/models/generator/javapojos/ApidocComments.scala
@@ -1,4 +1,4 @@
-package models.generator.android
+package models.generator.javapojos
 
 import lib.Constants
