+++ android-generator/src/test/scala/models/generator/javapojos/RetrofitUtilTest.scala
@@ -1,6 +1,7 @@
-package models.generator.android
+package models.generator.javapojos
 
 import org.scalatest.{Matchers, FlatSpec}
+import RetrofitUtil._
 
 class RetrofitUtilTest
   extends FlatSpec
